Move Group to separate file
